Even more file reorganization. Add cli option for starting cardano mode
