+++ backend/Dockerfile
@@ -18,5 +18,5 @@ ENV PATH="/root/.local/bin:$PATH"
 RUN poetry install --no-dev --no-interaction --no-ansi
 
 # Run the application as a Python module
-CMD ["poetry", "run", "python", "-m", "polydash", "-s", "settings.yaml"]
+CMD ["poetry", "run", "python", "-m", "polydash", "-s", "settings.yaml", "polygon"]
 # CMD ["poetry", "run", "python", "-m", "polydash"]
