--- backend/src/common/__init__.py
@@ -1,4 +1,0 @@
-class GetOrInsertMixin:
-    @classmethod
-    def get_or_insert(cls, **kwargs):
-        return cls.get(**kwargs) or cls(**kwargs)
