+++ backend/src/polydash/miners_ratings/model.py
@@ -4,8 +4,7 @@ from pony import orm
 from pony.orm import db_session
 from pydantic import BaseModel
 
-from polydash.common import GetOrInsertMixin
-from polydash.db import db
+from polydash.common.db import GetOrInsertMixin, db
 
 
 class MinerRiskHistory(db.Entity):
@@ -17,23 +16,10 @@ class MinerRiskHistory(db.Entity):
 
 
 class MinerRisk(db.Entity):
-    decay_coefficient = 0.9
-
     pubkey = orm.PrimaryKey(str)
     block_number = orm.Optional(int, index=True)
     numblocks = orm.Optional(int, default=0, index=True)
     risk = orm.Optional(float, default=0.0, index=True)
-
-    @classmethod
-    @db_session
-    def add_datapoint(cls, pubkey, num_risk_events, block_number):
-        datapoint = cls.get(pubkey=pubkey) or cls(pubkey=pubkey)
-        datapoint.risk = datapoint.risk * cls.decay_coefficient + num_risk_events
-        datapoint.numblocks += 1
-        datapoint.block_number = block_number
-        # Add historical record
-        MinerRiskHistory(pubkey=pubkey, block_number=block_number, risk=datapoint.risk, numblocks=datapoint.numblocks)
-        return datapoint
 
     @classmethod
     @db_session
