+++ backend/src/polydash/polygon/deanon/routes.py
@@ -1,6 +1,12 @@
-from fastapi import APIRouter
+from typing import List
 
-from polydash.deanon.model import *
+from fastapi import APIRouter
+from pony import orm
+from pony.orm import db_session
+from pydantic import BaseModel
+
+from polydash.polygon.deanon.model import DeanonNodeByBlockInDB, PeerToIPInDB, DeanonNodeByTxWithIP, DeanonNodeByTxInDB, \
+    PeerToIP, DeanonNodeByTx, DeanonNodeByBlock
 
 deanon_router = router = APIRouter(
     prefix="/deanon",
@@ -27,7 +33,7 @@ def get_list_of_deanon_nodes_by_txs_with_ips(nodes_from_db):
 
 @router.get("/by_txs")
 async def get_all_mappings_by_txs():
-    with orm.db_session:
+    with db_session:
         result = get_list_of_deanon_nodes_by_txs_with_ips(
             DeanonNodeByTx.select().sort_by(orm.desc(DeanonNodeByTx.confidence))
         )
@@ -36,7 +42,7 @@ async def get_all_mappings_by_txs():
 
 @router.get("/by_txs/by_pubkey/{pubkey}")
 async def get_by_txs_by_pubkey(pubkey: str):
-    with orm.db_session:
+    with db_session:
         result = get_list_of_deanon_nodes_by_txs_with_ips(
             DeanonNodeByTx.select(signer_key=pubkey).sort_by(
                 orm.desc(DeanonNodeByTx.confidence)
@@ -47,7 +53,7 @@ async def get_by_txs_by_pubkey(pubkey: str):
 
 @router.get("/by_txs/by_peer_id/{peer_id}")
 async def get_by_txs_by_peer_id(peer_id: str):
-    with orm.db_session:
+    with db_session:
         result = get_list_of_deanon_nodes_by_txs_with_ips(
             DeanonNodeByTx.select(peer_id=peer_id).sort_by(
                 orm.desc(DeanonNodeByTx.confidence)
@@ -78,7 +84,7 @@ def get_list_of_deanon_nodes_by_blocks_with_ips(nodes_from_db):
 
 @router.get("/by_blocks")
 async def get_all_mappings_by_blocks():
-    with orm.db_session:
+    with db_session:
         result = get_list_of_deanon_nodes_by_blocks_with_ips(
             DeanonNodeByBlock.select().sort_by(orm.desc(DeanonNodeByBlock.confidence))
         )
@@ -87,7 +93,7 @@ async def get_all_mappings_by_blocks():
 
 @router.get("/by_blocks/by_pubkey/{pubkey}")
 async def get_by_blocks_by_pubkey(pubkey: str):
-    with orm.db_session:
+    with db_session:
         result = get_list_of_deanon_nodes_by_blocks_with_ips(
             DeanonNodeByBlock.select(signer_key=pubkey).sort_by(
                 orm.desc(DeanonNodeByBlock.confidence)
@@ -98,7 +104,7 @@ async def get_by_blocks_by_pubkey(pubkey: str):
 
 @router.get("/by_blocks/by_peer_id/{peer_id}")
 async def get_by_blocks_by_peer_id(peer_id: str):
-    with orm.db_session:
+    with db_session:
         result = get_list_of_deanon_nodes_by_blocks_with_ips(
             DeanonNodeByBlock.select(peer_id=peer_id).sort_by(
                 orm.desc(DeanonNodeByBlock.confidence)
