+++ backend/src/polydash/polygon/w3router_watcher/settings.py
@@ -1,0 +1,5 @@
+from pydantic import BaseSettings
+
+
+class W3RouterSettings(BaseSettings):
+    w3_rpc_url: str = "http://localhost/rpc/update_nodes"
